--worked on the fixing the model issues

+++ user-service/src/main/java/com/mms/user/service/dtos/WalletRequestDTO.java
@@ -3,13 +3,15 @@ package com.mms.user.service.dtos;
 import com.mms.user.service.model.Wallet;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class WalletRequestDTO {
-    private long balance;
+    private BigDecimal balance;
     private Wallet.Status status;
     private int userId;
 }
